[feature] UserDetailsServiceImpl 서비스 단에 유지하기 위한 추상화
- 기존 구현 클래스였던 CustomUserDetailsService의 파일명을 AuthUserDetailsServiceImpl로 변경하고, 새로운 인터페이스를 CustomUserDetailsService로 분리함. 즉, AuthUserDetailsServiceImpl은 CustomUserDetailsService 인터페이스의 구현체임.
- common 모듈에서 CustomUserDetailsService 인터페이스에만 의존하도록 구성하여,
구현체인 AuthUserDetailsServiceImpl에 직접 의존하지 않음으로써 의존성 역전 원칙을 지킴.
- AuthUserDetailsServiceImpl은 기존처럼 user 도메인의 service 계층에 유지하면서,
User 도메인 전체를 common 모듈로 이동시키지 않아도 되는 구조가 됨.
