+++ module-api/src/main/java/com/welcommu/moduleapi/project/ProjectController.java
@@ -2,7 +2,7 @@ package com.welcommu.moduleapi.project;
 
 import com.welcommu.modulecommon.dto.ApiResponse;
 import com.welcommu.moduledomain.project.Project;
-import com.welcommu.moduledomain.user.CustomUserDetails;
+import com.welcommu.moduledomain.user.AuthUserDetailsImpl;
 import com.welcommu.moduleservice.project.ProjectService;
 import com.welcommu.moduleservice.project.dto.*;
 
@@ -34,7 +34,7 @@ public class ProjectController {
 
     @PostMapping
     public ResponseEntity<ApiResponse> createProject(
-        @AuthenticationPrincipal CustomUserDetails userDetails,
+        @AuthenticationPrincipal AuthUserDetailsImpl userDetails,
         @RequestBody ProjectCreateRequest dto
     ) {
