+++ module-common/src/main/java/com/welcommu/modulecommon/security/CustomUserDetailsService.java
@@ -1,27 +1,6 @@
 package com.welcommu.modulecommon.security;
-
-import com.welcommu.moduledomain.user.CustomUserDetails;
-import com.welcommu.moduledomain.user.User;
-import com.welcommu.modulerepository.user.UserRepository;
-import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
 
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        System.out.println("로그인 시도: " + email);  // 로그 추가
-
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + email));
-
-        return new CustomUserDetails(user);
-    }
+public interface CustomUserDetailsService {
+    UserDetails loadUserByUsername(String username);
 }
