+++ module-service/src/main/java/com/welcommu/moduleservice/auth/AuthUserDetailsServiceImpl.java
@@ -1,0 +1,27 @@
+package com.welcommu.moduleservice.auth;
+
+import com.welcommu.modulecommon.security.CustomUserDetailsService;
+import com.welcommu.moduledomain.user.AuthUserDetailsImpl;
+import com.welcommu.moduledomain.user.User;
+import com.welcommu.modulerepository.user.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AuthUserDetailsServiceImpl implements CustomUserDetailsService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        System.out.println("로그인 시도: " + email);  // 로그 추가
+
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + email));
+
+        return new AuthUserDetailsImpl(user);
+    }
+}
