--- module-service/src/main/java/com/welcommu/moduleservice/exception/ResourceNotFoundException.java
@@ -1,8 +1,0 @@
-package com.welcommu.moduleservice.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
